refactor RUT to use new transactional context
