+++ services/resource-usage-tracker/src/simcore_service_resource_usage_tracker/services/background_task_periodic_heartbeat_check.py
@@ -10,11 +10,12 @@ from models_library.resource_tracker import (
     ServiceRunStatus,
 )
 from pydantic import NonNegativeInt, PositiveInt
+from sqlalchemy.ext.asyncio import AsyncEngine
 
 from ..core.settings import ApplicationSettings
 from ..models.credit_transactions import CreditTransactionCreditsAndStatusUpdate
 from ..models.service_runs import ServiceRunStoppedAtUpdate
-from .modules.db.repositories.resource_tracker import ResourceTrackerRepository
+from .modules.db import credit_transactions_db, service_runs_db
 from .utils import compute_service_run_credit_costs, make_negative
 
 _logger = logging.getLogger(__name__)
@@ -23,7 +24,7 @@ _BATCH_SIZE = 20
 
 
 async def _check_service_heartbeat(
-    resource_tracker_repo: ResourceTrackerRepository,
+    db_engine: AsyncEngine,
     base_start_timestamp: datetime,
     resource_usage_tracker_missed_heartbeat_interval: timedelta,
     resource_usage_tracker_missed_heartbeat_counter_fail: NonNegativeInt,
@@ -55,7 +56,7 @@ async def _check_service_heartbeat(
                 missed_heartbeat_counter,
             )
             await _close_unhealthy_service(
-                resource_tracker_repo, service_run_id, base_start_timestamp
+                db_engine, service_run_id, base_start_timestamp
             )
         else:
             _logger.warning(
@@ -63,13 +64,16 @@ async def _check_service_heartbeat(
                 service_run_id,
                 missed_heartbeat_counter,
             )
-            await resource_tracker_repo.update_service_missed_heartbeat_counter(
-                service_run_id, last_heartbeat_at, missed_heartbeat_counter
+            await service_runs_db.update_service_missed_heartbeat_counter(
+                db_engine,
+                service_run_id=service_run_id,
+                last_heartbeat_at=last_heartbeat_at,
+                missed_heartbeat_counter=missed_heartbeat_counter,
             )
 
 
 async def _close_unhealthy_service(
-    resource_tracker_repo: ResourceTrackerRepository,
+    db_engine: AsyncEngine,
     service_run_id: ServiceRunId,
     base_start_timestamp: datetime,
 ):
@@ -80,8 +84,8 @@ async def _close_unhealthy_service(
         service_run_status=ServiceRunStatus.ERROR,
         service_run_status_msg="Service missed more heartbeats. It's considered unhealthy.",
     )
-    running_service = await resource_tracker_repo.update_service_run_stopped_at(
-        update_service_run_stopped_at
+    running_service = await service_runs_db.update_service_run_stopped_at(
+        db_engine, data=update_service_run_stopped_at
     )
 
     if running_service is None:
@@ -108,8 +112,8 @@ async def _close_unhealthy_service(
                 else CreditTransactionStatus.BILLED
             ),
         )
-        await resource_tracker_repo.update_credit_transaction_credits_and_status(
-            update_credit_transaction
+        await credit_transactions_db.update_credit_transaction_credits_and_status(
+            db_engine, data=update_credit_transaction
         )
 
 
@@ -118,19 +122,18 @@ async def periodic_check_of_running_services_task(app: FastAPI) -> None:
 
     # This check runs across all products
     app_settings: ApplicationSettings = app.state.settings
-    resource_tracker_repo: ResourceTrackerRepository = ResourceTrackerRepository(
-        db_engine=app.state.engine
-    )
+    _db_engine = app.state.engine
 
     base_start_timestamp = datetime.now(tz=timezone.utc)
 
     # Get all current running services (across all products)
-    total_count: PositiveInt = (
-        await resource_tracker_repo.total_service_runs_with_running_status_across_all_products()
+    total_count: PositiveInt = await service_runs_db.total_service_runs_with_running_status_across_all_products(
+        _db_engine
     )
 
     for offset in range(0, total_count, _BATCH_SIZE):
-        batch_check_services = await resource_tracker_repo.list_service_runs_with_running_status_across_all_products(
+        batch_check_services = await service_runs_db.list_service_runs_with_running_status_across_all_products(
+            _db_engine,
             offset=offset,
             limit=_BATCH_SIZE,
         )
@@ -138,7 +141,7 @@ async def periodic_check_of_running_services_task(app: FastAPI) -> None:
         await asyncio.gather(
             *(
                 _check_service_heartbeat(
-                    resource_tracker_repo=resource_tracker_repo,
+                    db_engine=_db_engine,
                     base_start_timestamp=base_start_timestamp,
                     resource_usage_tracker_missed_heartbeat_interval=app_settings.RESOURCE_USAGE_TRACKER_MISSED_HEARTBEAT_INTERVAL_SEC,
                     resource_usage_tracker_missed_heartbeat_counter_fail=app_settings.RESOURCE_USAGE_TRACKER_MISSED_HEARTBEAT_COUNTER_FAIL,
