--- services/resource-usage-tracker/src/simcore_service_resource_usage_tracker/services/modules/db/repositories/__init__.py
@@ -1,3 +1,0 @@
-from ._base import BaseRepository
-
-__all__: tuple[str, ...] = ("BaseRepository",)
