--- services/resource-usage-tracker/src/simcore_service_resource_usage_tracker/services/modules/db/repositories/_base.py
@@ -1,12 +1,0 @@
-from dataclasses import dataclass
-
-from sqlalchemy.ext.asyncio import AsyncEngine
-
-
-@dataclass
-class BaseRepository:
-    """
-    Repositories are pulled at every request
-    """
-
-    db_engine: AsyncEngine
