+++ services/resource-usage-tracker/src/simcore_service_resource_usage_tracker/services/pricing_units.py
@@ -11,21 +11,23 @@ from models_library.resource_tracker import (
     PricingUnitWithCostCreate,
     PricingUnitWithCostUpdate,
 )
+from sqlalchemy.ext.asyncio import AsyncEngine
 
-from ..api.rest.dependencies import get_repository
-from .modules.db.repositories.resource_tracker import ResourceTrackerRepository
+from ..api.rest.dependencies import get_resource_tracker_db_engine
+from .modules.db import pricing_plans_db
 
 
 async def get_pricing_unit(
     product_name: ProductName,
     pricing_plan_id: PricingPlanId,
     pricing_unit_id: PricingUnitId,
-    resource_tracker_repo: Annotated[
-        ResourceTrackerRepository, Depends(get_repository(ResourceTrackerRepository))
-    ],
+    db_engine: Annotated[AsyncEngine, Depends(get_resource_tracker_db_engine)],
 ) -> PricingUnitGet:
-    pricing_unit = await resource_tracker_repo.get_valid_pricing_unit(
-        product_name, pricing_plan_id, pricing_unit_id
+    pricing_unit = await pricing_plans_db.get_valid_pricing_unit(
+        db_engine,
+        product_name=product_name,
+        pricing_plan_id=pricing_plan_id,
+        pricing_unit_id=pricing_unit_id,
     )
 
     return PricingUnitGet(
@@ -42,21 +44,22 @@ async def get_pricing_unit(
 async def create_pricing_unit(
     product_name: ProductName,
     data: PricingUnitWithCostCreate,
-    resource_tracker_repo: Annotated[
-        ResourceTrackerRepository, Depends(get_repository(ResourceTrackerRepository))
-    ],
+    db_engine: Annotated[AsyncEngine, Depends(get_resource_tracker_db_engine)],
 ) -> PricingUnitGet:
     # Check whether pricing plan exists
-    pricing_plan_db = await resource_tracker_repo.get_pricing_plan(
-        product_name=product_name, pricing_plan_id=data.pricing_plan_id
+    pricing_plan_db = await pricing_plans_db.get_pricing_plan(
+        db_engine, product_name=product_name, pricing_plan_id=data.pricing_plan_id
     )
     # Create new pricing unit
-    pricing_unit_id, _ = await resource_tracker_repo.create_pricing_unit_with_cost(
-        data=data, pricing_plan_key=pricing_plan_db.pricing_plan_key
+    pricing_unit_id, _ = await pricing_plans_db.create_pricing_unit_with_cost(
+        db_engine, data=data, pricing_plan_key=pricing_plan_db.pricing_plan_key
     )
 
-    pricing_unit = await resource_tracker_repo.get_valid_pricing_unit(
-        product_name, data.pricing_plan_id, pricing_unit_id
+    pricing_unit = await pricing_plans_db.get_valid_pricing_unit(
+        db_engine,
+        product_name=product_name,
+        pricing_plan_id=data.pricing_plan_id,
+        pricing_unit_id=pricing_unit_id,
     )
     return PricingUnitGet(
         pricing_unit_id=pricing_unit.pricing_unit_id,
@@ -72,26 +75,30 @@ async def create_pricing_unit(
 async def update_pricing_unit(
     product_name: ProductName,
     data: PricingUnitWithCostUpdate,
-    resource_tracker_repo: Annotated[
-        ResourceTrackerRepository, Depends(get_repository(ResourceTrackerRepository))
-    ],
+    db_engine: Annotated[AsyncEngine, Depends(get_resource_tracker_db_engine)],
 ) -> PricingUnitGet:
     # Check whether pricing unit exists
-    await resource_tracker_repo.get_valid_pricing_unit(
-        product_name, data.pricing_plan_id, data.pricing_unit_id
+    await pricing_plans_db.get_valid_pricing_unit(
+        db_engine,
+        product_name=product_name,
+        pricing_plan_id=data.pricing_plan_id,
+        pricing_unit_id=data.pricing_unit_id,
     )
     # Get pricing plan
-    pricing_plan_db = await resource_tracker_repo.get_pricing_plan(
-        product_name, data.pricing_plan_id
+    pricing_plan_db = await pricing_plans_db.get_pricing_plan(
+        db_engine, product_name=product_name, pricing_plan_id=data.pricing_plan_id
     )
 
     # Update pricing unit and cost
-    await resource_tracker_repo.update_pricing_unit_with_cost(
-        data=data, pricing_plan_key=pricing_plan_db.pricing_plan_key
+    await pricing_plans_db.update_pricing_unit_with_cost(
+        db_engine, data=data, pricing_plan_key=pricing_plan_db.pricing_plan_key
     )
 
-    pricing_unit = await resource_tracker_repo.get_valid_pricing_unit(
-        product_name, data.pricing_plan_id, data.pricing_unit_id
+    pricing_unit = await pricing_plans_db.get_valid_pricing_unit(
+        db_engine,
+        product_name=product_name,
+        pricing_plan_id=data.pricing_plan_id,
+        pricing_unit_id=data.pricing_unit_id,
     )
     return PricingUnitGet(
         pricing_unit_id=pricing_unit.pricing_unit_id,
