+++ services/resource-usage-tracker/tests/unit/with_dbs/test_process_rabbitmq_message.py
@@ -8,9 +8,6 @@ from models_library.rabbitmq_messages import (
     SimcorePlatformStatus,
 )
 from servicelib.rabbitmq import RabbitMQClient
-from simcore_service_resource_usage_tracker.services.modules.db.repositories.resource_tracker import (
-    ResourceTrackerRepository,
-)
 from simcore_service_resource_usage_tracker.services.process_message_running_service import (
     _process_heartbeat_event,
     _process_start_event,
@@ -43,10 +40,7 @@ async def test_process_event_functions(
         pricing_unit_id=None,
         pricing_unit_cost_id=None,
     )
-    resource_tracker_repo: ResourceTrackerRepository = ResourceTrackerRepository(
-        db_engine=engine
-    )
-    await _process_start_event(resource_tracker_repo, msg, publisher)
+    await _process_start_event(engine, msg, publisher)
     output = await assert_service_runs_db_row(postgres_db, msg.service_run_id)
     assert output.stopped_at is None
     assert output.service_run_status == "RUNNING"
@@ -55,7 +49,7 @@ async def test_process_event_functions(
     heartbeat_msg = RabbitResourceTrackingHeartbeatMessage(
         service_run_id=msg.service_run_id, created_at=datetime.now(tz=timezone.utc)
     )
-    await _process_heartbeat_event(resource_tracker_repo, heartbeat_msg, publisher)
+    await _process_heartbeat_event(engine, heartbeat_msg, publisher)
     output = await assert_service_runs_db_row(postgres_db, msg.service_run_id)
     assert output.stopped_at is None
     assert output.service_run_status == "RUNNING"
@@ -66,7 +60,7 @@ async def test_process_event_functions(
         created_at=datetime.now(tz=timezone.utc),
         simcore_platform_status=SimcorePlatformStatus.OK,
     )
-    await _process_stop_event(resource_tracker_repo, stopped_msg, publisher)
+    await _process_stop_event(engine, stopped_msg, publisher)
     output = await assert_service_runs_db_row(postgres_db, msg.service_run_id)
     assert output.stopped_at is not None
     assert output.service_run_status == "SUCCESS"
